aris_backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
pandas==2.1.4
openpyxl==3.10.10
gunicorn==21.2.0
whitenoise==6.6.0
